third_party/WebKit/LayoutTests/fast/dom/DOMException/XPathException-expected.txt

Tests the properties of the exception thrown when using XPath.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS e.toString() is "TypeError: Type error"
PASS Object.prototype.toString.call(e) is "[object Error]"
PASS Object.prototype.toString.call(e.__proto__) is "[object Error]"
PASS e.constructor.toString() is "function TypeError() { [native code] }"
PASS e.constructor is window.TypeError
PASS successfullyParsed is true

TEST COMPLETE
